requirements.txt
asgiref==3.7.2
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2.1
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
google-api-core==2.15.0
google-api-python-client==2.113.0
google-auth==2.26.2
google-auth-httplib2==0.2.0
googleapis-common-protos==1.62.0
gunicorn==21.2.0
httplib2==0.22.0
idna==3.6
packaging==24.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyJWT==2.8.0
pyparsing==3.1.2
python-dotenv==1.0.0
pytz==2024.1
requests==2.31.0
rsa==4.9
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.1.0
